refactor: split package into core and pro
