+++ dev/kitchen-sink/Row3.tsx
@@ -1,9 +1,9 @@
-import { BoardRow } from '../../packages/react-components/src/BoardRow.js';
+import { BoardRow } from '../../packages/react-components-pro/src/BoardRow.js';
 import { Checkbox } from '../../packages/react-components/src/Checkbox.js';
 import { ComboBox } from '../../packages/react-components/src/ComboBox.js';
 import { ContextMenu } from '../../packages/react-components/src/ContextMenu.js';
-import { CookieConsent } from '../../packages/react-components/src/CookieConsent.js';
-import { Crud, crudPath } from '../../packages/react-components/src/Crud.js';
+import { CookieConsent } from '../../packages/react-components-pro/src/CookieConsent.js';
+import { Crud, crudPath } from '../../packages/react-components-pro/src/Crud.js';
 import { FormLayout } from '../../packages/react-components/src/FormLayout.js';
 import { Item } from '../../packages/react-components/src/Item.js';
 import { ListBox } from '../../packages/react-components/src/ListBox.js';
