+++ dev/pages/GridPro.tsx
@@ -1,5 +1,5 @@
-import { GridPro } from '../../packages/react-components/src/GridPro.js';
-import { GridProEditColumn } from '../../packages/react-components/src/GridProEditColumn.js';
+import { GridPro } from '../../packages/react-components-pro/src/GridPro.js';
+import { GridProEditColumn } from '../../packages/react-components-pro/src/GridProEditColumn.js';
 import { TextField } from '../../packages/react-components/src/TextField.js';
 
 type Item = {
