+++ packages/react-components-pro/README.md
@@ -1,0 +1,3 @@
+# Vaadin React components pro
+
+React wrappers for commercial Vaadin components.
