+++ packages/react-components-pro/package.json
@@ -1,0 +1,135 @@
+{
+  "name": "@vaadin/react-components-pro",
+  "version": "24.4.0-beta1",
+  "publishConfig": {
+    "access": "public"
+  },
+  "description": "React wrappers for commercial Vaadin components",
+  "type": "module",
+  "module": "index.js",
+  "main": "index.js",
+  "types": "index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "github.com/vaadin/react-components"
+  },
+  "scripts": {
+    "build": "npm run build:dev && npm run build:code && npm run build:update-packagejson",
+    "build:dev": "npm run build:generate",
+    "build:generate": "tsx ../../scripts/generator.ts",
+    "build:update-packagejson": "tsx ../../scripts/package-json-update.ts",
+    "build:code": "concurrently npm:build:code:*",
+    "build:code:ts": "tsx ../../scripts/build.ts",
+    "build:code:dts": "tsc -p tsconfig.build.json",
+    "build:code:copy-dts": "tsx ../../scripts/copy-dts.ts",
+    "version": "tsx ../../scripts/package-json-version.ts"
+  },
+  "dependencies": {
+    "@lit/react": "^1.0.0",
+    "@vaadin/board": "24.4.0-beta1",
+    "@vaadin/charts": "24.4.0-beta1",
+    "@vaadin/cookie-consent": "24.4.0-beta1",
+    "@vaadin/crud": "24.4.0-beta1",
+    "@vaadin/grid-pro": "24.4.0-beta1",
+    "@vaadin/map": "24.4.0-beta1",
+    "@vaadin/react-components": "24.4.0-beta1",
+    "@vaadin/rich-text-editor": "24.4.0-beta1"
+  },
+  "author": "Vaadin Ltd.",
+  "license": "SEE LICENSE IN LICENSE",
+  "peerDependencies": {
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "peerDependenciesMeta": {
+    "@types/react": {
+      "optional": true
+    },
+    "@types/react-dom": {
+      "optional": true
+    }
+  },
+  "sideEffects": false,
+  "files": [
+    "generated",
+    "renderers",
+    "utils",
+    "*.d.ts",
+    "*.d.ts.map",
+    "*.js",
+    "*.js.map"
+  ],
+  "exports": {
+    ".": {
+      "types": "./index.d.ts",
+      "default": "./index.js"
+    },
+    "./package.json": "./package.json",
+    "./Board.js": {
+      "types": "./Board.d.ts",
+      "default": "./Board.js"
+    },
+    "./BoardRow.js": {
+      "types": "./BoardRow.d.ts",
+      "default": "./BoardRow.js"
+    },
+    "./Chart.js": {
+      "types": "./Chart.d.ts",
+      "default": "./Chart.js"
+    },
+    "./ChartSeries.js": {
+      "types": "./ChartSeries.d.ts",
+      "default": "./ChartSeries.js"
+    },
+    "./CookieConsent.js": {
+      "types": "./CookieConsent.d.ts",
+      "default": "./CookieConsent.js"
+    },
+    "./Crud.js": {
+      "types": "./Crud.d.ts",
+      "default": "./Crud.js"
+    },
+    "./CrudEdit.js": {
+      "types": "./CrudEdit.d.ts",
+      "default": "./CrudEdit.js"
+    },
+    "./CrudEditColumn.js": {
+      "types": "./CrudEditColumn.d.ts",
+      "default": "./CrudEditColumn.js"
+    },
+    "./GridPro.js": {
+      "types": "./GridPro.d.ts",
+      "default": "./GridPro.js"
+    },
+    "./GridProEditColumn.js": {
+      "types": "./GridProEditColumn.d.ts",
+      "default": "./GridProEditColumn.js"
+    },
+    "./Map.js": {
+      "types": "./Map.d.ts",
+      "default": "./Map.js"
+    },
+    "./RichTextEditor.js": {
+      "types": "./RichTextEditor.d.ts",
+      "default": "./RichTextEditor.js"
+    },
+    "./Board": "./Board.js",
+    "./BoardRow": "./BoardRow.js",
+    "./Chart": "./Chart.js",
+    "./ChartSeries": "./ChartSeries.js",
+    "./CookieConsent": "./CookieConsent.js",
+    "./Crud": "./Crud.js",
+    "./CrudEdit": "./CrudEdit.js",
+    "./CrudEditColumn": "./CrudEditColumn.js",
+    "./GridPro": "./GridPro.js",
+    "./GridProEditColumn": "./GridProEditColumn.js",
+    "./Map": "./Map.js",
+    "./RichTextEditor": "./RichTextEditor.js",
+    "./utils/createComponent.d.ts": "./utils/createComponent.d.ts",
+    "./utils/createComponent.d.ts.map": "./utils/createComponent.d.ts.map",
+    "./utils/createComponent.js": "./utils/createComponent.js",
+    "./utils/createComponent.js.map": "./utils/createComponent.js.map"
+  }
+}
