+++ packages/react-components-pro/src/utils/createComponent.ts
@@ -1,0 +1,11 @@
+/**
+ * @license
+ * Copyright (c) 2000 - 2024 Vaadin Ltd.
+ *
+ * This program is available under Vaadin Commercial License and Service Terms.
+ *
+ *
+ * See https://vaadin.com/commercial-license-and-service-terms for the full
+ * license.
+ */
+export * from '@vaadin/react-components/utils/createComponent.js';
