+++ packages/react-components/README.md
@@ -1,3 +1,3 @@
 # Vaadin React components
 
-This package contains all React wrappers for Vaadin components (free and commercial).
+React wrappers for Vaadin components.
