+++ scripts/validate-build.ts
@@ -6,7 +6,8 @@ import { fileURLToPath } from 'node:url';
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const rootDir = resolve(__dirname, '..');
 const packagesDir = resolve(rootDir, 'packages');
-const componentsPackage = 'react-components';
+const corePackage = 'react-components';
+const proPackage = 'react-components-pro';
 
 /**
  * This function validates that the generated TypeScript definition file for a component
@@ -20,7 +21,7 @@ const componentsPackage = 'react-components';
  * @throws {Error} If the generated file contains any of the inherited properties.
  */
 async function validateInheritedProperties() {
-  const checkboxPath = resolve(packagesDir, componentsPackage, 'generated', 'Checkbox.d.ts');
+  const checkboxPath = resolve(packagesDir, corePackage, 'generated', 'Checkbox.d.ts');
 
   const typeDefinitionContent = await readFile(checkboxPath, 'utf8');
 
@@ -43,7 +44,7 @@ async function validateInheritedProperties() {
  * Validates that the generated directory exists.
  */
 async function hasGeneratedDir() {
-  [componentsPackage].forEach((packageName) => {
+  [proPackage, corePackage].forEach((packageName) => {
     if (!existsSync(resolve(packagesDir, packageName, 'generated'))) {
       throw new Error(`The generated directory does not exist. Run "npm run build" to generate it.`);
     }
@@ -51,11 +52,15 @@ async function hasGeneratedDir() {
 }
 
 /**
- * Validates that the package has css dir.
+ * Validates that the pro package does not have css dir and the core package does.
  */
 async function hasNoCssDir() {
-  if (!existsSync(resolve(packagesDir, componentsPackage, 'css'))) {
-    throw new Error(`The css directory does not exist in the components package.`);
+  if (existsSync(resolve(packagesDir, proPackage, 'css'))) {
+    throw new Error(`The css directory should not exist in the pro package.`);
+  }
+
+  if (!existsSync(resolve(packagesDir, corePackage, 'css'))) {
+    throw new Error(`The css directory does not exist in the core package.`);
   }
 }
 
@@ -63,7 +68,7 @@ async function hasNoCssDir() {
  * Validates that TS definition files have been copied (build:code:copy-dts has been run)
  */
 async function hasCopiedDts() {
-  if (!existsSync(resolve(packagesDir, componentsPackage, 'renderers', 'grid.d.ts'))) {
+  if (!existsSync(resolve(packagesDir, corePackage, 'renderers', 'grid.d.ts'))) {
     throw new Error(`The renderers/grid.d.ts file does not exist.`);
   }
 }
